modules/gamification/games/requirements.txt
# Games module dependencies
python-chess>=1.9.0
